Only use if explicitly set to LINES input.
