+++ pkg/s3select/json/reader.go
@@ -17,212 +17,117 @@
 package json
 
 import (
-	"bufio"
-	"bytes"
+	"errors"
 	"io"
-	"runtime"
 	"sync"
 
-	"github.com/bcicen/jstream"
-
 	"github.com/minio/minio/pkg/s3select/sql"
+
+	"github.com/bcicen/jstream"
 )
 
-// Reader - CSV record reader for S3Select.
+// Reader - JSON record reader for S3Select.
 type Reader struct {
-	args         *ReaderArgs
-	readCloser   io.ReadCloser    // raw input
-	buf          *bufio.Reader    // input to the splitter
-	columnNames  []string         // names of columns
-	nameIndexMap map[string]int64 // name to column index
-	current      []jstream.KVS    // current block of results to be returned
-	recordsRead  int              // number of records read in current slice
-	input        chan *queueItem  // input for workers
-	queue        chan *queueItem  // output from workers in order
-	err          error            // global error state, only touched by Reader.Read
-	bufferPool   sync.Pool        // pool of []byte objects for input
-	kvDstPool    sync.Pool        // pool of []jstream.KV used for output
-	close        chan struct{}    // used for shutting down the splitter before end of stream
-	readerWg     sync.WaitGroup   // used to keep track of async reader.
-}
-
-// queueItem is an item in the queue.
-type queueItem struct {
-	input []byte             // raw input sent to the worker
-	dst   chan []jstream.KVS // result of block decode
-	err   error              // any error encountered will be set here
+	args       *ReaderArgs
+	decoder    *jstream.Decoder
+	valueCh    chan *jstream.MetaValue
+	readCloser io.ReadCloser
 }
 
 // Read - reads single record.
-// Once Read is called the previous record should no longer be referenced.
 func (r *Reader) Read(dst sql.Record) (sql.Record, error) {
-	// If we have have any records left, return these before any error.
-	for len(r.current) <= r.recordsRead {
-		if r.err != nil {
-			return nil, r.err
+	v, ok := <-r.valueCh
+	if !ok {
+		if err := r.decoder.Err(); err != nil {
+			return nil, errJSONParsingError(err)
 		}
-		// Move to next block
-		item, ok := <-r.queue
-		if !ok {
-			r.err = io.EOF
-			return nil, r.err
-		}
-		//lint:ignore SA6002 Using pointer would allocate more since we would have to copy slice header before taking a pointer.
-		r.kvDstPool.Put(r.current)
-		r.current = <-item.dst
-		r.err = item.err
-		r.recordsRead = 0
+		return nil, io.EOF
 	}
-	kvRecord := r.current[r.recordsRead]
-	r.recordsRead++
+
+	var kvs jstream.KVS
+	if v.ValueType == jstream.Object {
+		// This is a JSON object type (that preserves key
+		// order)
+		kvs = v.Value.(jstream.KVS)
+	} else {
+		// To be AWS S3 compatible Select for JSON needs to
+		// output non-object JSON as single column value
+		// i.e. a map with `_1` as key and value as the
+		// non-object.
+		kvs = jstream.KVS{jstream.KV{Key: "_1", Value: v.Value}}
+	}
 
 	dstRec, ok := dst.(*Record)
 	if !ok {
 		dstRec = &Record{}
 	}
-	dstRec.KVS = kvRecord
+	dstRec.KVS = kvs
 	dstRec.SelectFormat = sql.SelectFmtJSON
 	return dstRec, nil
 }
 
 // Close - closes underlying reader.
 func (r *Reader) Close() error {
-	if r.close != nil {
-		close(r.close)
-		r.readerWg.Wait()
-		r.close = nil
+	// Close the input.
+	err := r.readCloser.Close()
+	for range r.valueCh {
+		// Drain values so we don't leak a goroutine.
+		// Since we have closed the input, it should fail rather quickly.
 	}
-	r.recordsRead = len(r.current)
-	if r.err == nil {
-		r.err = io.EOF
-	}
-	return r.readCloser.Close()
+	return err
 }
 
-// nextSplit will attempt to skip a number of bytes and
-// return the buffer until the next newline occurs.
-// The last block will be sent along with an io.EOF.
-func (r *Reader) nextSplit(skip int, dst []byte) ([]byte, error) {
-	if cap(dst) < skip {
-		dst = make([]byte, 0, skip+1024)
-	}
-	dst = dst[:skip]
-	if skip > 0 {
-		n, err := io.ReadFull(r.buf, dst)
-		if err != nil && err != io.ErrUnexpectedEOF {
-			// If an EOF happens after reading some but not all the bytes,
-			// ReadFull returns ErrUnexpectedEOF.
-			return dst[:n], err
-		}
-		dst = dst[:n]
-		if err == io.ErrUnexpectedEOF {
-			return dst, io.EOF
-		}
-	}
-	// Read until next line.
-	in, err := r.buf.ReadBytes('\n')
-	dst = append(dst, in...)
-	return dst, err
-}
-
-// csvSplitSize is the size of each block.
-// Blocks will read this much and find the first following newline.
-// 128KB appears to be a very reasonable default.
-const csvSplitSize = 128 << 10
-
-// startReaders will read the header if needed and spin up a parser
-// and a number of workers based on GOMAXPROCS.
-// If an error is returned no goroutines have been started and r.err will have been set.
-func (r *Reader) startReaders() {
-	r.bufferPool.New = func() interface{} {
-		return make([]byte, csvSplitSize+1024)
-	}
-
-	// Create queue
-	r.queue = make(chan *queueItem, runtime.GOMAXPROCS(0))
-	r.input = make(chan *queueItem, runtime.GOMAXPROCS(0))
-	r.readerWg.Add(1)
-
-	// Start splitter
-	go func() {
-		defer close(r.input)
-		defer close(r.queue)
-		defer r.readerWg.Done()
-		for {
-			next, err := r.nextSplit(csvSplitSize, r.bufferPool.Get().([]byte))
-			q := queueItem{
-				input: next,
-				dst:   make(chan []jstream.KVS, 1),
-				err:   err,
-			}
-			select {
-			case <-r.close:
-				return
-			case r.queue <- &q:
-			}
-
-			select {
-			case <-r.close:
-				return
-			case r.input <- &q:
-			}
-			if err != nil {
-				// Exit on any error.
-				return
-			}
-		}
-	}()
-
-	// Start parsers
-	for i := 0; i < runtime.GOMAXPROCS(0); i++ {
-		go func() {
-			for in := range r.input {
-				if len(in.input) == 0 {
-					in.dst <- nil
-					continue
-				}
-				dst, ok := r.kvDstPool.Get().([]jstream.KVS)
-				if !ok {
-					dst = make([]jstream.KVS, 0, 1000)
-				}
-
-				d := jstream.NewDecoder(bytes.NewBuffer(in.input), 0).ObjectAsKVS()
-				stream := d.Stream()
-				all := dst[:0]
-				for mv := range stream {
-					var kvs jstream.KVS
-					if mv.ValueType == jstream.Object {
-						// This is a JSON object type (that preserves key
-						// order)
-						kvs = mv.Value.(jstream.KVS)
-					} else {
-						// To be AWS S3 compatible Select for JSON needs to
-						// output non-object JSON as single column value
-						// i.e. a map with `_1` as key and value as the
-						// non-object.
-						kvs = jstream.KVS{jstream.KV{Key: "_1", Value: mv.Value}}
-					}
-					all = append(all, kvs)
-				}
-				// We don't need the input any more.
-				//lint:ignore SA6002 Using pointer would allocate more since we would have to copy slice header before taking a pointer.
-				r.bufferPool.Put(in.input)
-				in.input = nil
-				in.err = d.Err()
-				in.dst <- all
-			}
-		}()
-	}
-}
-
-// NewReader - creates new CSV reader using readCloser.
+// NewReader - creates new JSON reader using readCloser.
 func NewReader(readCloser io.ReadCloser, args *ReaderArgs) *Reader {
-	r := &Reader{
+	readCloser = &syncReadCloser{rc: readCloser}
+	d := jstream.NewDecoder(readCloser, 0).ObjectAsKVS()
+	return &Reader{
 		args:       args,
-		buf:        bufio.NewReaderSize(readCloser, csvSplitSize*2),
+		decoder:    d,
+		valueCh:    d.Stream(),
 		readCloser: readCloser,
-		close:      make(chan struct{}),
 	}
-	r.startReaders()
-	return r
+}
+
+// syncReadCloser will wrap a readcloser and make it safe to call Close
+// while reads are running.
+// All read errors are also postponed until Close is called and
+// io.EOF is returned instead.
+type syncReadCloser struct {
+	rc    io.ReadCloser
+	errMu sync.Mutex
+	err   error
+}
+
+func (pr *syncReadCloser) Read(p []byte) (n int, err error) {
+	// This ensures that Close will block until Read has completed.
+	// This allows another goroutine to close the reader.
+	pr.errMu.Lock()
+	defer pr.errMu.Unlock()
+	if pr.err != nil {
+		return 0, io.EOF
+	}
+	n, pr.err = pr.rc.Read(p)
+	if pr.err != nil {
+		// Translate any error into io.EOF, so we don't crash:
+		// https://github.com/bcicen/jstream/blob/master/scanner.go#L48
+		return n, io.EOF
+	}
+
+	return n, nil
+}
+
+var errClosed = errors.New("read after close")
+
+func (pr *syncReadCloser) Close() error {
+	pr.errMu.Lock()
+	defer pr.errMu.Unlock()
+	if pr.err == errClosed {
+		return nil
+	}
+	if pr.err != nil {
+		return pr.err
+	}
+	pr.err = errClosed
+	return pr.rc.Close()
 }
