+++ pkg/s3select/select.go
@@ -315,8 +315,11 @@ func (s3Select *S3Select) Open(getReader func(offset, length int64) (io.ReadClos
 			rc.Close()
 			return err
 		}
-
-		s3Select.recordReader = json.NewReader(s3Select.progressReader, &s3Select.Input.JSONArgs)
+		if s3Select.Input.JSONArgs.ContentType == "LINES" {
+			s3Select.recordReader = json.NewPReader(s3Select.progressReader, &s3Select.Input.JSONArgs)
+		} else {
+			s3Select.recordReader = json.NewReader(s3Select.progressReader, &s3Select.Input.JSONArgs)
+		}
 		return nil
 	case parquetFormat:
 		var err error
